core grpc things and grpc login/registration

+++ Android/app/src/main/java/ru/inwords/inwords/data/source/remote/session/TokenResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 class TokenResponse(
         @SerializedName("token") val accessToken: String = "",
-        @SerializedName("userId") val userId: Int = 0
+        @SerializedName("userId") val userId: Long = 0
 ) {
     val bearer: String get() = "Bearer $accessToken"
